update about screen
