+++ src/PenguinUpload/ClientApp/src/pages/Landing.vue
@@ -1,14 +1,6 @@
 <template>
   <div>
-    <div class="center">
-      <img src="public/icon.png" />
-      <h1 class="app-title">{{ appName }}</h1>
-    </div>
-    <div class="center title-subcontent">
-      <p>
-        a free, open file hosting service
-      </p>
-    </div>
+    <intro></intro>
     <div class="mg-top">
       <v-layout row>
         <v-flex xs12 md8 offset-md2>
@@ -27,6 +19,8 @@
 </template>
 
 <script>
+import intro from '../components/intro.vue'
+
 export default {
   data () {
     return {
@@ -40,16 +34,12 @@ export default {
       ]
     }
   },
-  computed: {
-    appName: function () {
-      return this.$store.state.data.appName
-    }
-  },
   methods: {
     visitUrl (u) {
       window.open(u)
     }
-  }
+  },
+  components: { intro }
 }
 </script>
